refactors kernel registration for GPU and CPU
