+++ tensorflow_addons/custom_ops/image/BUILD
@@ -45,5 +45,6 @@ custom_op_library(
         "cc/kernels/resampler_ops.h",
         "cc/kernels/sampling_functions.h",
         "cc/kernels/resampler_ops_gpu.cu.cc",
+        "cc/kernels/register_kernel_factory.h"
     ],
 )
